Fix deferred update problem which can break Berkeley DB databases.
Second and subsequent deferred update runs very likely have to update many
indicies for the incomplete segment.  If an index value refers to exactly
1 (one) record in the incomplete segment, adding any references to records
in the incomplete segment for the index value in a deferred update run will
break the index by adding a reference entry rather than replacing the
existing reference entry.  The added entry is correct and the original
entry can be deleted, but likely there are lots of them.

Module solentware_base.tools.berkeleydb_du_splice_fix repairs affected
databases.

The modules for Sqlite3 and other database engines are not affected.