Align FoundSet and RecordList with _DPTFoundSet and _DPTRecordList.
The FoundSet and RecordList classes are adjusted to the structure of the
_DPTFoundSet and _DPTRecordList classes.

The initialisation arguments for _DPTFoundSet and _DPTRecordList are
tidied up, and one property is defined to simplify the initialization
of _DPTFoundSet and _DPTRecordList instances.

Remove fieldname argument from _CursorDPT.__init__() and add fieldname as
argument to _CursorDPT.refresh_recordset_keep_position().