All deferred updates are done within transactions where available.
This is slower than before, especially with the DPT database engine, but
clients can arrange for interrupt and restart at transaction boundaries
which avoids taking a backup and repeating everything in some cases.

The code to do backups before deferred updates is removed.

Issue 4 is resolved.
Issue 5 is resolved.
Issue 6 is resolved.
Issue 7 is resolved.